Better type for pointer button mask
This is a very limited bit field, so use an 8 bit type to clearly show
how many bits are available.

+++ common/rfb/CMsgWriter.cxx
@@ -173,7 +173,7 @@ void CMsgWriter::writeKeyEvent(uint32_t keysym, uint32_t keycode, bool down)
 }
 
 
-void CMsgWriter::writePointerEvent(const Point& pos, int buttonMask)
+void CMsgWriter::writePointerEvent(const Point& pos, uint8_t buttonMask)
 {
   Point p(pos);
   if (p.x < 0) p.x = 0;
