+++ tests/unit/emulatemb.cxx
@@ -42,14 +42,14 @@ rfb::BoolParameter emulateMiddleButton("dummy_name", "dummy_desc", true);
 class TestClass : public EmulateMB
 {
 public:
-  void sendPointerEvent(const rfb::Point& pos, int buttonMask) override;
+  void sendPointerEvent(const rfb::Point& pos, uint8_t buttonMask) override;
 
-  struct PointerEventParams {rfb::Point pos; int mask; };
+  struct PointerEventParams {rfb::Point pos; uint8_t mask; };
 
   std::vector<PointerEventParams> results;
 };
 
-void TestClass::sendPointerEvent(const rfb::Point& pos, int buttonMask)
+void TestClass::sendPointerEvent(const rfb::Point& pos, uint8_t buttonMask)
 {
   PointerEventParams params;
   params.pos = pos;
