+++ vncviewer/EmulateMB.cxx
@@ -199,7 +199,7 @@ EmulateMB::EmulateMB()
 {
 }
 
-void EmulateMB::filterPointerEvent(const rfb::Point& pos, int buttonMask)
+void EmulateMB::filterPointerEvent(const rfb::Point& pos, uint8_t buttonMask)
 {
   int btstate;
   int action1, action2;
@@ -280,7 +280,7 @@ void EmulateMB::filterPointerEvent(const rfb::Point& pos, int buttonMask)
 void EmulateMB::handleTimeout(rfb::Timer *t)
 {
   int action1, action2;
-  int buttonMask;
+  uint8_t buttonMask;
 
   if (&timer != t)
     return;
@@ -312,7 +312,7 @@ void EmulateMB::handleTimeout(rfb::Timer *t)
   state = stateTab[state][4][2];
 }
 
-void EmulateMB::sendAction(const rfb::Point& pos, int buttonMask, int action)
+void EmulateMB::sendAction(const rfb::Point& pos, uint8_t buttonMask, int action)
 {
   assert(action != 0);
 
@@ -325,7 +325,7 @@ void EmulateMB::sendAction(const rfb::Point& pos, int buttonMask, int action)
   sendPointerEvent(pos, buttonMask);
 }
 
-int EmulateMB::createButtonMask(int buttonMask)
+int EmulateMB::createButtonMask(uint8_t buttonMask)
 {
   // Unset left and right buttons in the mask
   buttonMask &= ~0x5;
